Add a GPT DataSet class along with a datasetloader that by default creates the input-target strings based on BPE tokenizers
Add test code that shows how I can retrieve data in rows or in batches
where I control the batch size, the stride etc. The example output below
shows that the output - in Pytorch Tensors - is in batches with batch-size of 1
where the tensors are shifted by a stride of 1 to indicate input and prediction
targets.

batch 1:  [tensor([[  40,  367, 2885, 1464]]), tensor([[ 367, 2885, 1464, 1807]])]
batch 2:  [tensor([[ 367, 2885, 1464, 1807]]), tensor([[2885, 1464, 1807, 3619]])]