Fix a bug in attention scores calculation
If input is X rows then attention score of a query has to have X dimensions also,
as every element of input is dot-productized with the query (at an elemental level).

Thus if input has 6 rows then the attention score of a query (no matter how many dims it is represented with)
will also have 6 elements to it, with every element representing a "relation" between that query (input token)
and every other input token of that sentence. (for the LLM to get more understanding)

Of course the relationship calculation will later be based on trainable weights