Calculate context-vectors for all inputs using torch.dot
Creating a separate file (context_vec..) as I need to modify
the previous code in various places, to move from context-vector
generation for 1 input vector (query(2)) to all input tokens in the sentence.

Able to print out the 6 by 6 context vector array where every row represents
the attention scores (6) of that input token. There are some faults after that
since I haven't modified remainder of the code, and python doesn't offer me an easy
way to ignore a chunk of code in a file that already has multi-comments. (I could
use if False: etc but that's a pain with indenting that I have to redo later)