Implemetent causal masked self attention
We use tril of a 6 * 6 matrix (dim of attention scores) matrix of 1s below the
diagonal and 0s above (where the matrix represents token combos and upper diagonal
represents all future relationships of tokens) and multiply this lower diaginal unit
matrix with the original attention weights, to get a masked context matrix. Finally
we renormalize this matrix for every row so that the masked row elements all add up
to 1 in each row, thus showing how masked self attention is done.