Increase epochs to 20 to see the difference : post epoch 13 the output remains mostly the same, even though training loss goes down to the 0.1xx range. Val loss memories the training loss data and sticks to what it learn around epoch-2/epoch-3 range.
At the beginning of the training, both the training and validation set losses
sharply decrease, which is a sign that the model is learning. However, the training set loss
continues to decrease past the second epoch, whereas the validation loss stagnates.

This is a sign that the model is still learning, but it’s overfitting to the training set
past epoch 2.