Add model save and load function calls. Syntatically correct, need to run them Disable the load and let the save define the format (chicken and egg)
Now need to progressively enable code, fix compilation errors, and get training started
(checkpoints are enabled to progressively save and load state of the model on disk)

Added a TODO to also save the optimizer states of AdamW (along with the full model)
Optimizer states should be inspectable. Another TODO is getting GPU to work (my mac
is not an M series and metal is unsupported), and trying out distributed training
(which again implies a change of setup).

First goal is to get simple training running on Gutenberg books for my llm