modified some training params to make it more CPU friendly debugging (but note that once I scale these should go back to the defaults)
I'm also introducing bias (overfitting) by frequently stopping and starting
the training (as I am not saving the data-input state and basically restarting)
(This is evident in training_loss.png where loss increases and gets back to "normal"
once the previous repeated runs are crossed). Did not save the output (should add
my DualWriter code to this training loop)

I'm also not sure when the innermost 3rd for loop really ends (the input batch in target_batch)
as the value of input_batch (4) seems too small for target_batch (38k). (shdv saved the output)
Ran about 21 iterations of the first book (still in epoch 1). Need to debug if we
are converging out of epoch 1 (and some of the code from the previous training seems to be
different - again need to check)

The tokens_seen goes up much higher than the train_loader set. Again, need to verify if a batch
size of 4 (previously used) is the right granularity for this scale of training.