First clean run of gutenberg data. Ran for first 100 iterations (batches) of the first book. Stopped and saved the model. Results area already different from the main training. There is no overfitting now to the validation data as training_loss_100.png shows. The training loss itself widely fluctuates, that could be because the first set of data is not relevant at all to the query being asked.
The model is now set to continue running till it becomes as accurate as I'd
like it to be. Max iter for a book is currently set at 1000, I can remove it
also and have it run its 38K course. I did reduce Epochs to 3 because I don't
think I will need 10 epochs (also means that dropout is not going to be super
important from now on, which kind of mirrors what llama and gptOSS models are
also realizing in their architectures)