Load the model for the smaller training sample. Tested that the second iteration goes back to lower training losses (but the first iteration still starts at 10ish, which is really odd). Figure-1 and Figure-2 (.pngs) show the startling training vs validation losses and show the clear overfitting.
Now that gutenberg training has been launched, my intention is to add bells
and whistles first to this shorter training (including optimizer states) and
carry this over to gutenberg (because I don't see the covergence). Next I will
load up openAI weights in a separate folder.

I also need to enable an inference mode (this is training related) to check
the goodness of these chat models. And I also need to figure out how to use
a powerful GPU for training, as well as how to get to actual dstributed training
with powerful GPUs (in action) to actually complete the gutenberg training, and
then, some real stuff...