Add inference code and a log
Log shows a lot of things need to be fixed. Don't need the arch, don't
need an actual eval of a pre-built string. The following chat (not a loop
yet) shows that "the verdict" output prints gibberish (since its been
over-trained on one story only). But at least I have a way to check
inference independent of the model.