use the last token for prediction (based on how causal attention works) write up a prediction function and use it to predict accuracy for the 3 datasets
(had to comment out the train_loader data struct as I need to figure out a way to import it
from another file without executing any extra code). Once I do that ,the next step
is to determine the loss function that I will use for the training required for
classification. Also checking in 2 images that explain how causal attention means that
the last token actually holds the most info from the previous tokens, i.e. why I am
focusing on converting the last token to class prediction labels (as opposed to all of
the 4 rows of the output tensor)