Able to proceed with training for gtberg (100 batch run) of batch size 2
Back to CPU, on 80MB chunks locally on ubuntu, keeps budget to under 60%
of 16G mem. The logs and .png show that training loss fluctuates a bit
while validation loss is more constant. At this slow rate will *really*
take a long time to even finish one file, but model checkpoint saving
and loading is working (along with training states).

At this point both of my ubuntu laptops are operational for training
now for gutenberg on demosthenes (an activity that I had started but
not completed on my mac owing to the patience it needs)

Colab is also useful for Tesla T4 (the free tier) but it has its
limitations : I cannot guarantee time and sometimes the disconnects
happen randomly, like when I am sleeping, and the runtime instance
disconnects and I lose all of my data (not to mention the
mounting/upload/download cycle from google drive to colab and back to
downloads folder) An option is to use paid tier but the ROI should be
awesome.

I have also aquired more interesting datasets (wikipedia etc) so that
should be fun to try .... (other smaller trainings work and I have
instruction-follow also working. Can try with the full alpaca dataset)