continue training gtberg cpu 2 batchsize for 525 batches on ubuntu dell
the training loss distribution shows more downward spikes with training
loss. why does it keep coming back up to median? validation loss seems
consistent, and like my mac runs, I don't see convergence or a downward
motion. Is it because the text is all new and we're not repeating epochs
on the same data (like in 1 file), is it because the file is still too
big (80mb each).

Anyways I continue to make progress ..