Implement cosine decay along with learning rate warmup
Do both in the training loop. Show that the curve for LR first
increases to peak within 20% of the cycles and the remainder of
the 80% is a downward half cycle cosine curve. This is to prevent
overshooting of minima especially as the training stabilizes towards
the later stages of training.